Add optimization for wasm
